Remove the 'processing' state for jobs
It is a very volatile state, intended only to show something during the
(very very short) time between a file is added to the app and it starts
being read. The message is barely readable because in the best case it
is only shown for a few milliseconds, so it does not make sense.