Decouple Job class from Presenter in main thread
Currently, Job objects are storing their identifiers, which is a piece
of information only needed because of the current business logic inside
the Presenter. As such, Job objects should not be aware of this piece of
information.

This has been solved by abusing the Job registry Map() in the Presenter,
to have a reverse mapping between the Job objects and their identifiers,
resulting in a much cleaner and decoupled code.