Decouple UI messages from job states
Until now, the type used for representing job states (a Symbol) allowed
to include a string as description, and this was abused to put there the
corresponding UI message.

As a side effect, constants related to states have been renamed so they
are more readable and provide a bit of context when used.