Fix bug fired by early errors
If an error happens too early in the 'booting' process of the app, extra
errors will appear because some elements disappeared, etc.

This is perfectly normal and does not really need any fix, because the
app is halted when this happens, but it does not look good and it's very
easy to fix.