Use new channel interface with service worker
Previously there was no communication with the service worker except for
the lifecycle events, but now there will be, for a start for getting the
application version code. In order to achieve this the new communication
interface will be used with the service worker, too.

Please note that for the time being the service worker is not aware of
this, it won't be accepting commands or sending back replies. The new
system is only being set up for the future.