Demodularize the service worker
Apparently, Firefox is not going to support module-type service workers
for the time being. At least, not soon enough. It is the last "big" web
browser not supporting this, and I suppose they are very proud of being
so different, but the fact is the app would not work on Firefox and it
MUST work on this web browser.

So, the service worker has been demodularized for the time being.