Change version handling
After a lot (really, A LOT) of different approaches for version handling
in the application, and a lot of different definitions of what comprises
a certain version of the application, a (hopefully) definitive solution
has been adopted.

A version of the application is a specific set of fetchable resources,
that is, assets in the application code terminology, which are stored in
the application cache managed by the service worker.

As such, the version is handled by the service worker, which informs the
main code about it when it is controlling the fetching of resources. The
service worker uses the channel communication mechanism which was added
previously to refactor message passing.

The version code has been improved.